+++ Cargo.toml
@@ -80,7 +80,7 @@ tracing = ["vmm/tracing", "tracer/tracing"]
 members = [
     "api_client",
     "arch",
-    "block_util",
+    "block",
     "devices",
     "event_monitor",
     "hypervisor",
@@ -89,12 +89,10 @@ members = [
     "option_parser",
     "pci",
     "performance-metrics",
-    "qcow",
     "rate_limiter",
     "serial_buffer",
     "test_infra",
     "tracer",
-    "vhdx",
     "vhost_user_block",
     "vhost_user_net",
     "virtio-devices",
