+++ block/Cargo.toml
@@ -1,26 +1,27 @@
 [package]
-name = "block_util"
+name = "block"
 version = "0.1.0"
-authors = ["The Cloud Hypervisor Authors"]
 edition = "2021"
+authors = ["The Cloud Hypervisor Authors", "The Chromium OS Authors"]
 
 [features]
 default = []
 io_uring = ["dep:io-uring"]
 
 [dependencies]
+byteorder = "1.4.3"
+crc32c = "0.6.3"
 io-uring = { version = "0.5.13", optional = true }
 libc = "0.2.139"
 log = "0.4.17"
-qcow = { path = "../qcow" }
+remain = "0.2.6"
 smallvec = "1.10.0"
 thiserror = "1.0.40"
+uuid = { version = "1.3.4", features = ["v4"] }
 versionize = "0.1.10"
 versionize_derive = "0.1.4"
-vhdx = { path = "../vhdx" }
 virtio-bindings = { version = "0.2.0", features = ["virtio-v5_0_0"] }
 virtio-queue = "0.8.0"
 vm-memory = { version = "0.11.0", features = ["backend-mmap", "backend-atomic", "backend-bitmap"] }
 vm-virtio = { path = "../vm-virtio" }
 vmm-sys-util = "0.11.0"
-
