+++ block/src/qcow/mod.rs
@@ -2,17 +2,16 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE-BSD-3-Clause file.
 
-#[macro_use]
-extern crate log;
-
 mod qcow_raw_file;
 mod raw_file;
 mod refcount;
 mod vec_cache;
 
-use crate::qcow_raw_file::QcowRawFile;
-use crate::refcount::RefCount;
-use crate::vec_cache::{CacheMap, Cacheable, VecCache};
+use crate::qcow::{
+    qcow_raw_file::QcowRawFile,
+    refcount::RefCount,
+    vec_cache::{CacheMap, Cacheable, VecCache},
+};
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
 use libc::{EINVAL, ENOSPC, ENOTSUP};
 use remain::sorted;
@@ -25,7 +24,7 @@ use vmm_sys_util::{
     write_zeroes::WriteZeroesAt,
 };
 
-pub use crate::raw_file::RawFile;
+pub use crate::qcow::raw_file::RawFile;
 
 #[sorted]
 #[derive(Debug)]
@@ -362,8 +361,8 @@ fn max_refcount_clusters(refcount_order: u32, cluster_size: u32, num_clusters: u
 /// # Example
 ///
 /// ```
+/// # use block::qcow::{self, QcowFile, RawFile};
 /// # use std::io::{Read, Seek, SeekFrom};
-/// # use qcow::{self, QcowFile, RawFile};
 /// # fn test(file: std::fs::File) -> std::io::Result<()> {
 ///     let mut raw_img = RawFile::new(file, false);
 ///     let mut q = QcowFile::from(raw_img).expect("Can't open qcow file");
