+++ block/src/qcow/refcount.rs
@@ -7,8 +7,10 @@ use std::io;
 
 use libc::EINVAL;
 
-use crate::qcow_raw_file::QcowRawFile;
-use crate::vec_cache::{CacheMap, Cacheable, VecCache};
+use crate::qcow::{
+    qcow_raw_file::QcowRawFile,
+    vec_cache::{CacheMap, Cacheable, VecCache},
+};
 
 #[derive(Debug)]
 pub enum Error {
