+++ block/src/qcow_sync.rs
@@ -1,63 +1,64 @@
 // Copyright © 2021 Intel Corporation
 //
-// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: Apache-2.0 AND BSD-3-Clause
 
 use crate::async_io::{AsyncIo, AsyncIoResult, DiskFile, DiskFileError, DiskFileResult};
+use crate::qcow::{QcowFile, RawFile, Result as QcowResult};
 use crate::AsyncAdaptor;
 use std::collections::VecDeque;
 use std::fs::File;
+use std::io::{Seek, SeekFrom};
 use std::sync::{Arc, Mutex, MutexGuard};
-use vhdx::vhdx::{Result as VhdxResult, Vhdx};
 use vmm_sys_util::eventfd::EventFd;
 
-pub struct VhdxDiskSync {
-    vhdx_file: Arc<Mutex<Vhdx>>,
+pub struct QcowDiskSync {
+    qcow_file: Arc<Mutex<QcowFile>>,
 }
 
-impl VhdxDiskSync {
-    pub fn new(f: File) -> VhdxResult<Self> {
-        Ok(VhdxDiskSync {
-            vhdx_file: Arc::new(Mutex::new(Vhdx::new(f)?)),
+impl QcowDiskSync {
+    pub fn new(file: File, direct_io: bool) -> QcowResult<Self> {
+        Ok(QcowDiskSync {
+            qcow_file: Arc::new(Mutex::new(QcowFile::from(RawFile::new(file, direct_io))?)),
         })
     }
 }
 
-impl DiskFile for VhdxDiskSync {
+impl DiskFile for QcowDiskSync {
     fn size(&mut self) -> DiskFileResult<u64> {
-        Ok(self.vhdx_file.lock().unwrap().virtual_disk_size())
+        let mut file = self.qcow_file.lock().unwrap();
+
+        file.seek(SeekFrom::End(0)).map_err(DiskFileError::Size)
     }
 
     fn new_async_io(&self, _ring_depth: u32) -> DiskFileResult<Box<dyn AsyncIo>> {
-        Ok(
-            Box::new(VhdxSync::new(self.vhdx_file.clone()).map_err(DiskFileError::NewAsyncIo)?)
-                as Box<dyn AsyncIo>,
-        )
+        Ok(Box::new(QcowSync::new(self.qcow_file.clone())) as Box<dyn AsyncIo>)
     }
 }
 
-pub struct VhdxSync {
-    vhdx_file: Arc<Mutex<Vhdx>>,
+pub struct QcowSync {
+    qcow_file: Arc<Mutex<QcowFile>>,
     eventfd: EventFd,
     completion_list: VecDeque<(u64, i32)>,
 }
 
-impl VhdxSync {
-    pub fn new(vhdx_file: Arc<Mutex<Vhdx>>) -> std::io::Result<Self> {
-        Ok(VhdxSync {
-            vhdx_file,
-            eventfd: EventFd::new(libc::EFD_NONBLOCK)?,
+impl QcowSync {
+    pub fn new(qcow_file: Arc<Mutex<QcowFile>>) -> Self {
+        QcowSync {
+            qcow_file,
+            eventfd: EventFd::new(libc::EFD_NONBLOCK)
+                .expect("Failed creating EventFd for QcowSync"),
             completion_list: VecDeque::new(),
-        })
+        }
     }
 }
 
-impl AsyncAdaptor<Vhdx> for Arc<Mutex<Vhdx>> {
-    fn file(&mut self) -> MutexGuard<Vhdx> {
+impl AsyncAdaptor<QcowFile> for Arc<Mutex<QcowFile>> {
+    fn file(&mut self) -> MutexGuard<QcowFile> {
         self.lock().unwrap()
     }
 }
 
-impl AsyncIo for VhdxSync {
+impl AsyncIo for QcowSync {
     fn notifier(&self) -> &EventFd {
         &self.eventfd
     }
@@ -68,7 +69,7 @@ impl AsyncIo for VhdxSync {
         iovecs: &[libc::iovec],
         user_data: u64,
     ) -> AsyncIoResult<()> {
-        self.vhdx_file.read_vectored_sync(
+        self.qcow_file.read_vectored_sync(
             offset,
             iovecs,
             user_data,
@@ -83,7 +84,7 @@ impl AsyncIo for VhdxSync {
         iovecs: &[libc::iovec],
         user_data: u64,
     ) -> AsyncIoResult<()> {
-        self.vhdx_file.write_vectored_sync(
+        self.qcow_file.write_vectored_sync(
             offset,
             iovecs,
             user_data,
@@ -93,7 +94,7 @@ impl AsyncIo for VhdxSync {
     }
 
     fn fsync(&mut self, user_data: Option<u64>) -> AsyncIoResult<()> {
-        self.vhdx_file
+        self.qcow_file
             .fsync_sync(user_data, &self.eventfd, &mut self.completion_list)
     }
