+++ fuzz/Cargo.toml
@@ -9,7 +9,7 @@ edition = "2021"
 cargo-fuzz = true
 
 [dependencies]
-block_util = { path = "../block_util" }
+block = { path = "../block" }
 devices = { path = "../devices" }
 epoll = "4.3.1"
 libc = "0.2.147"
@@ -18,9 +18,7 @@ linux-loader = { version = "0.9.0", features = ["elf", "bzimage", "pe"] }
 micro_http = { git = "https://github.com/firecracker-microvm/micro-http", branch = "main" }
 net_util = { path = "../net_util" }
 once_cell = "1.18.0"
-qcow = { path = "../qcow" }
 seccompiler = "0.3.0"
-vhdx = { path = "../vhdx" }
 virtio-devices = { path = "../virtio-devices" }
 virtio-queue = "0.8.0"
 vmm = { path = "../vmm" }
