+++ vhost_user_block/Cargo.toml
@@ -7,13 +7,12 @@ build = "../build.rs"
 
 [dependencies]
 argh = "0.1.9"
-block_util = { path = "../block_util" }
+block = { path = "../block" }
 env_logger = "0.10.0"
 epoll = "4.3.3"
 libc = "0.2.139"
 log = "0.4.17"
 option_parser = { path = "../option_parser" }
-qcow = { path = "../qcow" }
 vhost = { version = "0.7.0", features = ["vhost-user-slave"] }
 vhost-user-backend = "0.9.0"
 virtio-bindings = "0.2.0"
