+++ virtio-devices/Cargo.toml
@@ -10,7 +10,7 @@ default = []
 [dependencies]
 anyhow = "1.0.71"
 arc-swap = "1.5.1"
-block_util = { path = "../block_util" }
+block = { path = "../block" }
 byteorder = "1.4.3"
 epoll = "4.3.3"
 event_monitor = { path = "../event_monitor" }
